CY-3678 Use JSON for the profile context file (#1151)
* ProfileContext: a json-based implementation

Instead of storing the ProfileContext as a yaml object, store
it as a json dict.
Also implement storing and loading to/from a dict.

I'm gonna (for now?) keep the "name" field in the dict, even though
python-side it's called `.profile_name`.

Also add a method for loading the context from the (old) yaml
context.
Not adding any unittests for the yaml stuff because tbh I'm not gonna
be able to make them better than just testing by hand now: it would
have to be tied to a hardcoded yaml string, and obviously it can't
test the general form of the yaml files out there.

* Drop suppress_errors where possible and minor refactors

- drop suppress_errors in functions that are well-suited for, or
  even already are, returning a None
- handle both context.json and context for the `is_profile_exists` check
- use try/except IOError instead of isfile for the get_active_profile
  (EAFP is better than LBYL wrt. possible race conditions)
- drop a few unittests that were specifically testing for suppress_errors
- drop the specialcased local profile in get_profile_context (and a test)
  because instead, we just don't call it for local profiles instead.
- env.delete_profile silently Just Deletes™ and does nothing if it's
  already deleted. This is because env.delete_profile is also used
  internally (eg. `cfy profiles set --profile-name`). For the case
  where we _do_ want to throw an error (`cfy prof delete` on a
  nonexistent profile), throw the error from inside the cli command
  itself. (also made it correctly return nonzero in that case,
  and updated a unittest to reflect that)

* Update paths in unittests

- for the filename, instead of context, use context.json
- the original CLOUDIFY_WORKDIR includes .cloudify so make the test
  one include it as well. This makes it possible to run a test in
  a debugger, and use the test profiles via `cfy` at the same time,
  which makes for easier debugging

* Rename ssh_key_path to ssh_key

The original object's attribute was .ssh_key, so this is keeping
the same name.
Before, it only changed in to_dict, which affects the `cfy prof li`
display. We renamed it everywhere to ssh_key, so rename it there as
well.

* Drop the constant

I don't want the constant for the filename, because that encourages
importing it and accessing the file by other means than the
one load method we have. I don't want the profile to be loaded
using any other methods.

* Drop pyyaml-specific test

* Add a unittest

load an old-style profile, update it, save it, reload it and
check that the updated values are present in the reloaded profile

+++ cloudify_cli/commands/profiles.py
@@ -41,7 +41,7 @@ from ..commands.cluster import update_profile_logic as update_cluster_profile
 EXPORTED_KEYS_DIRNAME = '.exported-ssh-keys'
 EXPORTED_SSH_KEYS_DIR = os.path.join(env.PROFILES_DIR, EXPORTED_KEYS_DIRNAME)
 PROFILE_COLUMNS = ['name', 'manager_ip', 'manager_username', 'manager_tenant',
-                   'ssh_user', 'ssh_key_path', 'ssh_port', 'kerberos_env',
+                   'ssh_user', 'ssh_key', 'ssh_port', 'kerberos_env',
                    'rest_port', 'rest_protocol', 'rest_certificate']
 CLUSTER_PROFILE_COLUMNS = PROFILE_COLUMNS[:1] + ['hostname', 'host_ip'] \
     + PROFILE_COLUMNS[2:]
@@ -298,11 +298,11 @@ def delete(profile_name, logger):
     `PROFILE_NAME` is the IP of the manager the profile manages.
     """
     logger.info('Deleting profile {0}...'.format(profile_name))
-    try:
-        env.delete_profile(profile_name)
-        logger.info('Profile deleted')
-    except CloudifyCliError as ex:
-        logger.info(str(ex))
+    if not env.is_profile_exists(profile_name):
+        raise CloudifyCliError('Profile {0} does not exist'
+                               .format(profile_name))
+    env.delete_profile(profile_name)
+    logger.info('Profile deleted')
 
 
 def set_profile(profile_name,
